+++ lib/invoices.ts
@@ -19,7 +19,7 @@ import { models } from './models'
 
 import { Orm } from './orm'
 
-import { v4 } from 'uuid'
+import * as shortid from 'shortid';
 
 import { createPaymentOptions } from './invoice'
 
@@ -172,21 +172,63 @@ export class Invoice extends Orm {
 
 }
 
-export async function createInvoice(params: NewInvoice): Promise<Invoice> {
+interface CreateInvoice {
+  account: Account,
+  amount: number;
+  currency?: string;
+  fee_rate_level?: string;
+  redirect_url?: string;
+  webhook_url?: string;
+  wordpress_site_url?: string;
+  external_id?: string;
+  memo?: string;
+  email?: string;
+  business_id?: string;
+  location_id?: string;
+  register_id?: string;
+}
 
-  const uid = v4();
 
-  var newInvoice: any = {
 
-    denomination_currency: params.account.denomination,
+export async function createInvoice(params: CreateInvoice): Promise<Invoice> {
 
-    denomination_amount: params.amount,
+  const uid = shortid.generate();
 
-    currency: params.account.denomination,
+  var { webhook_url, account, amount, currency } = params
 
-    amount: params.amount,
+  if (!webhook_url) {
 
-    account_id: params.account.id,
+    webhook_url = account.get('webhook_url')
+
+  }
+
+  const newInvoice: any = {
+
+    denomination_currency: currency || account.denomination,
+
+    denomination_amount: amount,
+
+    currency: currency || account.denomination,
+
+    amount: amount,
+
+    account_id: account.id,
+
+    webhook_url,
+
+    external_id: params.external_id,
+
+    business_id: params.business_id,
+
+    location_id: params.location_id,
+
+    register_id: params.register_id,
+
+    memo: params.memo,
+
+    wordpress_site_url: params.wordpress_site_url,
+
+    fee_rate_level: params.fee_rate_level,
 
     status: 'unpaid',
 
@@ -202,57 +244,15 @@ export async function createInvoice(params: NewInvoice): Promise<Invoice> {
 
   }
 
-  if (params.webhook_url) {
-
-    if (validator.isURL(params.webhook_url, {protocols: 'https'})) {
-
-      newInvoice['webhook_url'] = params.webhook_url
-
-    } else {
-
-      throw new InvalidWebhookURL(params.webhook_url)
-
-    }
-
-  } else {
-
-    newInvoice['webhook_url'] = `${config.get('API_BASE')}/v1/api/test/webhooks`
-
-  }
-
   var record = await models.Invoice.create(newInvoice);
-
-  const account = await findAccount(record.account_id)
-
-  var webhook_url = params.webhook_url
-
-  if (!webhook_url) {
-
-    webhook_url = account.get('webhook_url')
-
-  }
-
-  if (webhook_url) {
-
-    await createWebhook({
-
-      invoice_uid: uid,
-
-      url: webhook_url,
-
-      account_id: record.account_id
-
-    })
-  }
 
   let invoice = new Invoice(record)
 
+  await createWebhook(invoice)
+
   const options = await createPaymentOptions(account.record, invoice)
 
-  await log.info('invoice.created', {
-    invoice_uid: invoice.uid,
-    account_id: invoice.account_id
-  })
+  log.info('invoice.created', record.toJSON())
 
   return invoice;