+++ lib/webhooks.ts
@@ -59,13 +59,21 @@ export async function sendWebhookForInvoice(invoiceUid: string, type: string = '
 
     } catch(e) {
 
-      response_code = e.response.statusCode;
+      log.debug('webhook.error', e)
+      log.debug('webhook.error', e.message)
 
-      if (typeof e.response.body !== 'string') {
-        response_body = JSON.stringify(e.response.text); 
-      } else {
-        response_body = e.response.body; 
+      if (e.response) {
+
+        response_code = e.response.statusCode;
+
+        if (typeof e.response.body !== 'string') {
+          response_body = JSON.stringify(e.response.text); 
+        } else {
+          response_body = e.response.body; 
+        }
+
       }
+
       error = e.message;
 
       ended_at = new Date();
@@ -73,8 +81,6 @@ export async function sendWebhookForInvoice(invoiceUid: string, type: string = '
       status = 'failed'
 
     }
-
-
 
     let webhook = await models.Webhook.create({
       account_id: invoice.account_id,
@@ -245,9 +251,18 @@ interface CreateWebhook {
   account_id: number;
 }
 
-export async function createWebhook(where: CreateWebhook): Promise<Webhook> {
+export async function createWebhook(invoice: Invoice): Promise<Webhook> {
 
-  let [webhook] = await models.Webhook.findOrCreate({where, defaults: where})
+  const where = {
+    invoice_uid: invoice.uid,
+    url: invoice.get('webhook_url') || DEFAULT_WEBHOOK_URL,
+    account_id: invoice.get('account_id')
+  }
+
+  let [webhook] = await models.Webhook.findOrCreate({
+    where,
+    defaults: where
+  })
 
   if (!webhook) { throw new WebhookNotFound() }