+++ plugins/dash/index.ts
@@ -7,8 +7,6 @@ import { fromSatoshis, Payment } from '../../lib/pay'
 
 import { any } from 'bluebird'
 
-import {generateInvoice} from '../../lib/invoice';
-
 import {log, models} from '../../lib'
 
 import { getTransaction } from './lib/jsonrpc';
@@ -16,6 +14,7 @@ import { getTransaction } from './lib/jsonrpc';
 export { getTransaction }
 
 import * as blockcypher from '../../lib/blockcypher'
+
 import * as blockchair from '../../lib/blockchair'
 
 import * as http from 'superagent';