+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.2.6](https://github.com/anypay/anypay/compare/v1.2.5...v1.2.6) (2022-07-29)
+
+
+### Bug Fixes
+
+* **webhook-url:** actually set webhook url for invoice ([#1038](https://github.com/anypay/anypay/issues/1038)) ([4ae4d41](https://github.com/anypay/anypay/commit/4ae4d412e0d7d4a80bc3ea11ef6880670c614001))
+
+## [1.2.5](https://github.com/anypay/anypay/compare/v1.2.4...v1.2.5) (2022-07-28)
+
+
+### Bug Fixes
+
+* **readd-payment-options:** in POST /invoices API response, return payment_options ([#1035](https://github.com/anypay/anypay/issues/1035)) ([276c398](https://github.com/anypay/anypay/commit/276c3980a3637afe307080678e70be403a8777a7))
+
+## [1.2.4](https://github.com/anypay/anypay/compare/v1.2.3...v1.2.4) (2022-07-28)
+
+
+### Bug Fixes
+
+* **failed-webhooks-bug:** fix bug when failed webhooks don't log to database ([#1033](https://github.com/anypay/anypay/issues/1033)) ([ee927f3](https://github.com/anypay/anypay/commit/ee927f3c9c30dce78f29d60f4c97263189cfd122))
+
 ## [1.2.3](https://github.com/anypay/anypay/compare/v1.2.2...v1.2.3) (2022-07-26)